Move thumbnail logic into separate component `Thumbnail`
